+++ pkg/scheduler/partition.go
@@ -500,6 +500,7 @@ func (pc *PartitionContext) getQueueInternal(name string) *objects.Queue {
 // GetPartitionQueues builds the queue info for the whole queue structure to pass to the webservice
 func (pc *PartitionContext) GetPartitionQueues() dao.PartitionQueueDAOInfo {
 	partitionQueueDAOInfo := pc.root.GetPartitionQueueDAOInfo(true)
+	partitionQueueDAOInfo.Partition = common.GetPartitionNameWithoutClusterID(pc.Name)
 	partitionQueueDAOInfo.PartitionID = pc.ID
 	return partitionQueueDAOInfo
 }
